+++ plugins/dates/src/02-ranges/index.js
@@ -1,14 +1,14 @@
 const parseDate = require('../03-parse')
-const parseInterval = require('./intervals')
+const repeating = require('./00-repeats')
 const ranges = [].concat(require('./01-two-times'), require('./02-two-date'), require('./03-one-date'))
 
 // loop thru each range template
 const parseRange = function (doc, context) {
   // parse-out 'every week ..'
-  let interval = parseInterval(doc, context) || {}
+  let repeats = repeating(doc, context) || {}
   // if it's *only* an interval response
   if (doc.found === false) {
-    return Object.assign({}, interval, { start: null, end: null })
+    return Object.assign({}, repeats, { start: null, end: null })
   }
   // try each template in order
   for (let i = 0; i < ranges.length; i += 1) {
@@ -23,7 +23,7 @@ const parseRange = function (doc, context) {
       }
       let res = fmt.parse(m, context)
       if (res !== null) {
-        return Object.assign({}, interval, res)
+        return Object.assign({}, repeats, res)
       }
     }
   }
@@ -44,7 +44,7 @@ const parseRange = function (doc, context) {
       unit: unit.setTime ? 'time' : unit.unit,
     }
   }
-  let combined = Object.assign({}, interval, res)
+  let combined = Object.assign({}, repeats, res)
   // ensure start is not after end
   // console.log(combined)
   if (combined.start && combined.end && combined.start.d.epoch > combined.end.d.epoch) {