+++ plugins/dates/tests/interval.test.js
@@ -15,6 +15,7 @@ test('test first generated-date', function (t) {
     ['every hour tomorrow', '2021-02-17T08:00:00.000+05:30'],
     ['any month this year', '2021-01-01T08:00:00.000+05:30'],
     ['any month next year', '2022-01-01T08:00:00.000+05:30'],
+    ['every thursday at 2pm', '2021-02-18T14:00:00.000+05:30'],
   ]
   arr.forEach((a) => {
     let doc = nlp(a[0])