+++ src/02-tagger/03-contractions/06-ranges.js
@@ -1,5 +1,6 @@
 const isRange = /^([0-9.]{1,3}[a-z]{0,2})[-–—]([0-9]{1,3}[a-z]{0,2})$/i
 const timeRange = /^([0-9][0-9]?(:[0-9][0-9])?(am|pm)?)[-–—]([0-9][0-9]?(:[0-9][0-9])?(am|pm)?)$/i
+
 //split '2-4' into '2 to 4'
 const checkRange = function (term) {
   if (term.tags.PhoneNumber === true) {